+++ activesupport/lib/active_support/core_ext/big_decimal/conversions.rb
@@ -1,15 +1,7 @@
 require 'bigdecimal'
 require 'bigdecimal/util'
-require 'yaml'
 
 class BigDecimal
-  YAML_MAPPING = { 'Infinity' => '.Inf', '-Infinity' => '-.Inf', 'NaN' => '.NaN' }
-
-  def encode_with(coder)
-    string = to_s
-    coder.represent_scalar(nil, YAML_MAPPING[string] || string)
-  end
-
   # Backport this method if it doesn't exist
   unless method_defined?(:to_d)
     def to_d
